Fixed another serialization clash
